Modularize active chat assets to avoid merge conflicts
