+++ public/assets/css/chat.css
@@ -1,0 +1,228 @@
+.chat-shell {
+    display: grid;
+    gap: 1.25rem;
+}
+
+.chat-noscript {
+    padding: 0.75rem 1rem;
+    border-radius: 0.75rem;
+    border: 1px dashed rgba(148, 163, 184, 0.4);
+    background: rgba(15, 23, 42, 0.5);
+    font-family: 'Roboto Mono', monospace;
+    font-size: 0.85rem;
+    letter-spacing: 0.06em;
+}
+
+.chat-window {
+    min-height: clamp(16rem, 24vw, 22rem);
+    max-height: 24rem;
+    border-radius: 1rem;
+    border: 1px solid rgba(0, 255, 136, 0.25);
+    background: rgba(0, 15, 8, 0.65);
+    backdrop-filter: blur(6px);
+    padding: 1.1rem 1.25rem;
+    overflow-y: auto;
+    box-shadow: inset 0 0 30px rgba(0, 255, 136, 0.08);
+    display: flex;
+    flex-direction: column;
+    gap: 0.9rem;
+}
+
+.chat-window::-webkit-scrollbar {
+    width: 0.35rem;
+}
+
+.chat-window::-webkit-scrollbar-thumb {
+    background: rgba(0, 255, 136, 0.45);
+    border-radius: 9999px;
+}
+
+.chat-meta {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 1rem;
+}
+
+.chat-status {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.6rem;
+    font-family: 'Roboto Mono', monospace;
+    font-size: 0.85rem;
+    letter-spacing: 0.06em;
+    color: rgba(148, 163, 184, 0.7);
+}
+
+.chat-status-indicator {
+    width: 0.7rem;
+    height: 0.7rem;
+    border-radius: 9999px;
+    display: inline-block;
+    background: rgba(148, 163, 184, 0.6);
+    box-shadow: 0 0 6px rgba(148, 163, 184, 0.3);
+}
+
+.chat-clear {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.4rem;
+    border: 1px solid rgba(0, 255, 136, 0.3);
+    border-radius: 0.65rem;
+    padding: 0.4rem 0.8rem;
+    background: rgba(15, 23, 42, 0.5);
+    color: rgba(204, 255, 204, 0.85);
+    font-family: 'Roboto Mono', monospace;
+    font-size: 0.8rem;
+    letter-spacing: 0.05em;
+    cursor: pointer;
+    transition: transform 0.2s ease, border-color 0.2s ease, box-shadow 0.2s ease;
+}
+
+.chat-clear:hover {
+    transform: translateY(-1px);
+    border-color: rgba(0, 255, 136, 0.6);
+    box-shadow: 0 0 12px rgba(0, 255, 136, 0.25);
+}
+
+.chat-entry {
+    display: grid;
+    gap: 0.65rem;
+}
+
+.chat-label {
+    font-family: 'Roboto Mono', monospace;
+    font-size: 0.8rem;
+    letter-spacing: 0.06em;
+    color: rgba(204, 255, 204, 0.8);
+}
+
+.chat-controls {
+    display: flex;
+    gap: 0.75rem;
+}
+
+.chat-controls input[type='text'] {
+    flex: 1 1 auto;
+    border-radius: 0.75rem;
+    border: 1px solid rgba(0, 255, 136, 0.35);
+    background: rgba(0, 0, 0, 0.45);
+    color: rgba(224, 255, 224, 0.95);
+    padding: 0.75rem 1rem;
+    font-family: 'Roboto Mono', monospace;
+    letter-spacing: 0.04em;
+    font-size: 0.9rem;
+    transition: border-color 0.2s ease, box-shadow 0.2s ease;
+}
+
+.chat-controls input[type='text']::placeholder {
+    color: rgba(226, 232, 240, 0.45);
+}
+
+.chat-controls input[type='text']:focus {
+    outline: none;
+    border-color: var(--color-neon-primary);
+    box-shadow: 0 0 12px rgba(0, 255, 136, 0.25);
+}
+
+.chat-send {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.4rem;
+    padding: 0.75rem 1.2rem;
+    border-radius: 0.75rem;
+    background: linear-gradient(90deg, rgba(0, 255, 136, 0.9), rgba(59, 130, 246, 0.75));
+    border: none;
+    color: #020617;
+    font-weight: 700;
+    text-transform: uppercase;
+    letter-spacing: 0.08em;
+    cursor: pointer;
+    box-shadow: 0 12px 25px rgba(0, 255, 136, 0.2);
+    transition: transform 0.2s ease, box-shadow 0.2s ease;
+}
+
+.chat-send i,
+.chat-clear i {
+    width: 1rem;
+    height: 1rem;
+}
+
+.chat-send:hover {
+    transform: translateY(-1px);
+    box-shadow: 0 16px 32px rgba(0, 255, 136, 0.3);
+}
+
+.chat-message {
+    display: grid;
+    gap: 0.35rem;
+    padding: 0.75rem 0.9rem;
+    border-radius: 0.75rem;
+    border: 1px solid rgba(0, 255, 136, 0.18);
+    background: rgba(4, 16, 12, 0.75);
+    font-family: 'Roboto Mono', monospace;
+}
+
+.chat-message--user {
+    margin-left: auto;
+    border-color: rgba(59, 130, 246, 0.4);
+    background: rgba(30, 64, 175, 0.25);
+    box-shadow: 0 0 16px rgba(59, 130, 246, 0.15);
+}
+
+.chat-message--ally {
+    margin-right: auto;
+    border-color: rgba(34, 197, 94, 0.45);
+    background: rgba(6, 78, 59, 0.35);
+    box-shadow: 0 0 16px rgba(34, 197, 94, 0.18);
+}
+
+.chat-message--system {
+    margin: 0 auto;
+    border-style: dashed;
+    border-color: rgba(148, 163, 184, 0.35);
+    color: rgba(226, 232, 240, 0.85);
+    background: rgba(15, 23, 42, 0.45);
+}
+
+.chat-author {
+    font-size: 0.75rem;
+    letter-spacing: 0.08em;
+    text-transform: uppercase;
+    color: rgba(148, 163, 184, 0.75);
+}
+
+.chat-message p {
+    margin: 0;
+    font-size: 0.9rem;
+    line-height: 1.5;
+    color: rgba(226, 255, 226, 0.9);
+}
+
+.chat-message time {
+    font-size: 0.65rem;
+    letter-spacing: 0.08em;
+    text-transform: uppercase;
+    color: rgba(148, 163, 184, 0.6);
+    justify-self: flex-end;
+}
+
+.chat-shell button {
+    letter-spacing: 0.15em;
+}
+
+@media (max-width: 720px) {
+    .chat-meta {
+        flex-direction: column;
+        align-items: stretch;
+    }
+
+    .chat-controls {
+        flex-direction: column;
+    }
+
+    .chat-send {
+        width: 100%;
+    }
+}
